+++ test/serialize.js
@@ -19,7 +19,7 @@ describe("serialize PG", () => {
   it("special examples", () => {
     const examples = {
       "1 :; a:b x:\"y z\" x:y: z": [{ id: "1", labels: [";"], properties: { a: ["b"], x: ["y z"], "x:y": ["z"] } }],
-      "1:2 :[] )(:\",\"": [{ id: "1:2", labels: ["[]"], properties: { ")(":[","] } }],
+      "1:2 :\"[]\" )(:\",\"": [{ id: "1:2", labels: ["[]"], properties: { ")(":[","] } }],
       "": [{ id: "" }],
       x: [{ id: "x", properties: { y: null } }],
       "x y:1": [{ id: "x", properties: { y: [{},1,null] } }],